requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
sqlparse==0.4.2
psycopg2==2.9.3
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
djangorestframework-simplejwt==5.1.0
django-s3-storage==0.13.6
# boto3==1.18.58
# botocore==1.26.5 #1.23.54
# six<=1.15.0
ffmpeg==1.4
ffmpeg-python==0.2.0
pillow==2.2.1
requests==2.26.0
fcm-django==1.0.12
firebase-admin==5.2.0
